Add example YAML and field descriptions
The content and format here is based off of the great example of the
cert-policy-controller. The example is from the `policy-pod` in the
policy-collection.

The fields are from configurationpolicy_types.go, and I marked them as
required or optional depending on if they seemed necessary for the
policy to work as expected; most of them had `omitempty` in their go
struct tags, so they aren't *technically* required.

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>