Add test for whitespace in list
Currently, if an item in a list has whitespace at the beginning or end,
it won't get matched by the same item because only one of them is
trimmed. As a result, and enforce policy with an item like that will
keep appending duplicates of the item on the list indefinitely. This
test demonstrates the issue with a Deployment.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>