Fix an issue that caused integer comparison to fail
This fixes the situation where the controller is comparing what is in
the policy versus what already exists, and the object being compared
contains an integer.

The code in the mergeSpecs function marshals and unmarshals the existing
object and the object in the policy. Presumably this is to create a copy
of the objects before modification. By doing so with the encoding/json
package, it converts any integers to float64.

This caused an issue in the checkFieldsWithSort function which compared
the existing object with the merged object, since the existing object
referenced here did not go through the marshal and unmarshal process,
but the merged object did. This meant that the integers in the merged
object were of type float64, but they remained integers in the existing
object. When checkFieldsWithSort converts both integers to strings to
compare, if the one of type float64 is a large enough number, it gets
printed in scientific notation where as the one that remained of type
integer does not. This would cause a mismatch even though it should have
matched.

The solution is quite simple. Just start using the json package from
Kubernetes so that the same unmarshaling process occurs in all cases.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>