Protect the policy cache from unintended updates (#167)
Deep copy the policy before passing it to handleObjectTemplates since even
though handleObjectTemplates accepts a copy (i.e. not a pointer) of the
policy, policy.Spec.ObjectTemplates is a slice of pointers, so any
modifications to the objects in that slice will be reflected in
the PolicyMap cache, which can have unintended side effects.

This manifested itself when a policy has templates which when resolved,
still have intentional template delimiters (i.e. {{ and }}) in the result.
On the first time being processed, everything would go well. On the
second or third time, when the policy cache was not updated again by the
reconciler, the policy in the cache already had resolved templates. This
would lead to the templates code trying to reprocess the already
resolved policy.

Additionally to this, this caused some updates to referenced resource
objects in templates to not get properly updated in the policy.

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=2007575

Signed-off-by: mprahl <mprahl@users.noreply.github.com>

Co-authored-by: mprahl <mprahl@users.noreply.github.com>