Wait for kubectl commands in tests to complete
The kubectl function used by the tests was not checking if there was an
error during the execution, which could cause tests to continue past
failing commands. This will now wait for it to complete, and print the
output of the command if it fails, to make it easier to debug.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>