Remove global recorder and reconcilingAgent
Functions that used these globals now became methods on the reconciler.

For the recorder in particular, it was confusing which one was being
used, and whether it had been correctly initialized. It is now easier
to see where it comes from (in main.go) and allows some "if nil" logic
to be removed.

I removed the unit test that would have needed re-writing for this
because there is an e2e test that tests that behavior more thoroughly.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>