Disable caching of Secrets (#222)
The controller-runtime caches objects by default when retrieved
using `r.Get` and this creates a watch on the GVK based on the
`WATCH_NAMESPACE` value we use. This undesirable for the `r.Get` call
for the `policy-encryption-key` secret since this is a value that rarely
changes and caching is handled in the controller logic. This avoids an
unnecessary watch on the Kubernetes API.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>