Simplify the logic that determines which policies to process
The `Reconcile` method populated a shared map of `ConfigurationPolicy`
of objects. This added an entry in the map for every namespace that the
`namespaceSelector` applied to, however, the `namespaceSelector` only
applies to the object templates, so this was unnecessary. To compensate
for that, the `PeriodicallyExecConfigPolicies` would "flatten" the map
to deduplicate the entries.

To simplify this, the `PeriodicallyExecConfigPolicies` method just
directly queries the controller-runtime cache for all `ConfigurationPolicy`
objects, which would contain all the `ConfigurationPolicy` objects in the
watched namespace(s).

This improves performance, simplifies the code, and reduces some of the
cases where the `ConfigurationPolicy` object was updated but the update
wasn't yet picked up by the `PeriodicallyExecConfigPolicies` method.

This will help for the implementation of adding a custom evaluation
interval per policy.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>