Add a configurable evaluation interval for configuration policies
This adds the field `spec.evaluationInterval` which is a map that
accepts the keys `compliant` and `noncompliant`. Each value is a
duration for the minimum wait time between policy evaluations. The
values can also be set to `never` to disable evaluating the policy
entirely after a certain compliance state is reached. In this case, the
compliance message mentions that it will not be reevaluated.

On the technical implementation side, two new status fields are added.
The first is `status.lastEvaluated` which is a RFC3339/ISO-8601
formatted timestamp. This keeps track of when the policy was last
evaluated for the controller to know when the next evaluation should
take place. It is also useful information for the user.

The second status field that is added is
`status.lastEvaluatedGeneration`, which is the generation of the
`ConfigurationPolicy` object when the policy was last evaluated. This is
for the controller to know when to trigger an immediate evaluation when
the `ConfigurationPolicy` object's `spec` is updated.

Because these two status fields need to be set on every evaluation, the
status is always updated per evaluation, however, an event is not sent
when it's just these two status fields that were updated.

Resolves:
https://github.com/stolostron/backlog/issues/20006

Signed-off-by: mprahl <mprahl@users.noreply.github.com>