Cache the API discovery information
Prior to this commit, the API discovery information was refreshed on
every cycle (i.e. every 10 seconds) but this puts unnecessary load on
the Kubernetes API.

This commit makes it so that the API discovery information is refreshed
whenever an API resource (not an object) that is not in the discovery
information is encountered. It is also periodically updated every 10
minutes.

Resolves:
https://github.com/stolostron/backlog/issues/20382

Signed-off-by: mprahl <mprahl@users.noreply.github.com>