Handle objects that have omitted empty values
If a policy specifies an empty array or map in its object definition
and the property is set to `omitEmpty`, then the controller will
mark the policy as noncompliant when it compares the policy with the
object retrieved from the API since the property wouldn't be returned.

This commit makes it so that an omitted array or map value is equal to
it being empty.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>