Refresh the discovery info only once per object template (#245)
This handles the case where `isObjectNamespaced` refreshes the discovery
info but cannot find the API resource, so it is marked as non
namespaced. Then `getMapping` refreshes the discovery info and finds the
API resource, but it was previously marked as not namespaced.

Additionally, this speeds things up when processing policies related to
objects where the API resource is not installed.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>