Allow an alternative kubeconfig for evaluating policies
This allows the controller to run on a separate cluster than the one it
is managing.

As part of this, global variables were removed to make it simpler to
determine where the Kubernetes client and configuration is coming from.

Related:
https://github.com/stolostron/backlog/issues/24361

Signed-off-by: mprahl <mprahl@users.noreply.github.com>
(cherry picked from commit 263f9b94e5042ea4707d7991a7a1012f6d0f56d4)