Fix: update UID when recreating an object
When an object was re-created by the policy after being deleted, the
UID in the object's properties was not updated. Then, the logic for
DeleteIfCreated pruning would see the UID as stale and not remove the
object, even though it *was* created by the policy.

Refs:
 - https://github.com/stolostron/backlog/issues/24677

Co-authored-by: Will Kutler <wkutler@redhat.com>
Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 0a20010cfa756983d03c49d8ff3314b711e97750)