Update named objects in arrays instead of appending
In the case of a Deployment object's container array, if a policy
updates the image of a container, the container should be updated.

Before this commit, a new container with the same name would be
erroneously appended to the containers array and cause the Kubernetes
API update request to fail and make the policy non-compliant.

This assumption is safe due to the Kubernetes API conventions:
https://github.com/kubernetes/community/blob/master/contributors/devel/sig-architecture/api-conventions.md#lists-of-named-subobjects-preferred-over-maps

Relates:
https://github.com/stolostron/backlog/issues/25061
https://bugzilla.redhat.com/show_bug.cgi?id=2117728

Signed-off-by: mprahl <mprahl@users.noreply.github.com>
(cherry picked from commit 5d93fa56090d464f4992b3032c7c6cdbaf1d7758)