Evaluate policies on every loop when evaluationInterval isn't set
This reverts back to the original behavior before evaluationInterval was
added.

Before this commit, the default value was 10s, which was supposed to be
the equivalent of every loop. In practice, this often ended up being
every other loop when there were few policies.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>
(cherry picked from commit 2cdedb9702305537b5478127dc3005ce20477009)