Fix  ConfigurationPolicies are left behind bug
Description of problem:
The pruneObjectBehavior of a ConfigurationPolicy does not help clean up objects when the policy is edited such that the managed object's name is changed. Ideally, editing the policy would result in deleting the old object, if the pruneObjectBehavior field was set.

Signed-off-by: Yi Rae Kim <yikim@redhat.com>
(cherry picked from commit d3603379037f6e48a11077753e5d5331213cdbeb)