Reduce the copying when evaluating a policy
This reduces the recursive copying to just the value in the object
template since that is what is used for merging the existing value in.
For safety, the existing object that is used to merge into the template
object is copied just once up front.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>
(cherry picked from commit d877c5e176a4c2245bdd2ed48b79cd27c8acc0bd)