Fix compliance when created resource has a status
Previously, the template would always be marked as compliant after the
resource it defines was created. But if the object definition included
a `status`, then the object might not actually match what was desired.
Now, in this case, the template will be marked as non-compliant.

Refs:
 - https://issues.redhat.com/browse/ACM-7020

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 04a91c0b0d46ae1645564101be8006d146d28088)