Reduce some noisy logs
Because this controller runs on a loop (is not event-driven), it will
often log messages that it is not doing anything. These are not often
very useful, but can add up to a huge amount of text over a period of
time. These changes generally decrease the number of logs this process
will produce.

Refs:
 - https://issues.redhat.com/browse/ACM-7778

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 8d561ad749ccf927479e6dcd353c4dc79ae9fbe1)