Retry policy evaluation when object changed during evaluation
This is important for the dry-run to not send a noncompliant error
because the dry-run failed due to an object conflict. On the normal
update case, it saves the time it takes for the policy to be evaluated
again.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>
(cherry picked from commit d2a4e0436a062e1028a0125b5ceb87ea5beb5625)