Add a default namespace for operators
This will allow OperatorPolicies to use a default namespace when the
user does not specify one in the policy. For example, on OpenShift, it's
conventional to install operators to the openshift-operators namespace
when they are cluster-wide. This default namespace is set via a command
line flag, so it can be customized per deployment.

This commit also adds a validation condition to OperatorPolicy, for
reporting when the subscription or operatorgroup specs are invalid.

Refs:
 - https://issues.redhat.com/browse/ACM-9896

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 71bd52227cb108a8d7161b071c40823325fb88c3)