Add more validation to the OperatorPolicy
The `ValidPolicySpec` condition now reports when the subscription or
operatorGroup in the policy have unknown fields. It also reports if the
InstallPlanApproval value in the subscription is not correct.

That condition also reports when the namespace for the subscription
does not exist, and the controller now watches that namespace so that if
it is created (or deleted), the policy will be reconciled and the status
will be updated.

Refs:
 - https://issues.redhat.com/browse/ACM-9993

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit c4cfa083012ff35b0bdc85a3f46ab24b27cfb994)