Relax event requirement for a flaky test
When the subscription has a ConstraintsNotSatisfiable condition, the
exact content of the message is not always consistent. We have added
other things to sort the message better, but it looks like we missed
this test which was still assuming a certain part was first.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 12f0324e60912ebbbe14708f899c4e84a50046bd)