Relax event requirement for a flaky test
When the subscription has a ConstraintsNotSatisfiable condition, the
exact content of the message is not always consistent. We have added
other things to sort the message better, but it looks like we missed
this test which was still assuming a certain part was first.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit fa5b2358a9c01bc3b09f3e189ef8b2b9dde030d5)

Cherry-pick for:
 - https://issues.redhat.com/browse/ACM-11190