Wait for actual ConstraintsNotSatisfiable reason
This test started failing relatively often. Debugging revealed that the
Subscription was getting into a ResolutionFailed=True condition, but
that it did not have the expected ConstraintsNotSatisfiable reason; it
looked like it was failing to resolve because the connection was refused
to the catalog.

Refs:
 - https://issues.redhat.com/browse/ACM-11414

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit d02b4b97f67738e2809b7b8a36bc0548164e868d)