Prevent subscription creation if opgroup incorrect
Previously, the subscription would be created by an enforced musthave
OperatorPolicy regardless of the status of the OperatorGroup. But this
could lead to a bad result where the operator is running in the wrong
mode because it is not using the correct operatorgroup. While the policy
was correctly non-compliant in this case, it should not create the
subscription.

Refs:
 - https://issues.redhat.com/browse/ACM-11507

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 414cd18c62dfd4075f11ff5d43d2d55c338837d9)