Simplify condition for ConstraintsNotSatisfiable
Previously, the controller needed to parse and sort the message
associated with the ConstraintsNotSatisfiable condition on the
Subscription, because the order of certain parts ot the message were
inconsistent. The intent was to not update the OperatorPolicy condition
when the actual situation on the cluster was unchanged.

In some situations, particularly when the subscription is deleted and
quickly recreated, the condition on the Subscription does not just have
parts in an inconsistent order: it can alternate between two different
clauses. To prevent constant updates to the OperatorPolicy status, it
now just reports a generic "conditions not satisfiable" condition, and
directs the user to the Subscription for more information. This also
reduces noise in the compliance events be removing the lists of operator
versions.

Refs:
 - https://issues.redhat.com/browse/ACM-11453

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 9c9d42a13e74b15d73bdcd35129d19e68de260d8)