Use suggested namespace in package if not set
Previously, if the namespace was not set for a Subscription in an
OperatorPolicy, the default namespace on the controller (if set) would
be used in all cases. But many packages specify a suggested namespace in
their PackageManifests, which would be better to use, when provided.

Additionally, if the OperatorPolicy was already managing a subscription,
the namespace of that subscription will be used (if not set in the
policy). This may prevent some issues that could occur if the suggested
namespace for an operator changes at some point.

Refs:
 - https://issues.redhat.com/browse/ACM-12057

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 125c471ed9fddc0674d7493d4c2bcc4d6ebbf23a)