Add recordInfoEvent method
This new method uses the reconciler's recorder to emit an info event.
The same action was being done in-line in several places, which have
been updated to use this method. The benefits are that we could change
this function to be synchronous and emit an error more easily in the
future (like we've done with the compliance event) and that it is more
clear what the intent/requirement of the event is now.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 6eae8ba89ecb05e5d4ecf819f06d15c73329beb8)