Initialize namespace selections in constructor
Previously, this map was initialized when being setup with a controller
runtime manager. It is not necessary to wait until then, and it was
preventing any possibility of running the reconcilers without a real
manager (for example, in tests).

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit a98f4aaeae5ba0d8b9a135877613a65a2b8ca69c)