Add GenerateMappings command
This would allow a user to save a set of REST mappings from an existing
cluster which the config-policy-controller reconciler *could* use
instead of a "live" discovery API. Currently the setup in main.go for
the controller *does not* connect this, but it would be possible to add
in the future.

The included "default-mappings.yaml" file is the output of the tool from
a simple kind cluster.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit d914d301ab450a8b2bb08eb926928d449ccdcc15)