Add dryrun command
This command allows for the testing of configuration policies without a
cluster. "Cluster state" is provided via input files, which are
simulated in a fake cluster. A real config-policy-controller reconciler
is run against this fake cluster, and the resulting policy status, as
well as the compliance events, are captured and can be displayed to the
user. If the policy is noncompliant, the command exits with error code 2
(to distinguish it from other possible exit reasons).

This also exposes the GenerateMappings command as a subcommand.

Refs:
 - https://issues.redhat.com/browse/ACM-14161

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 6e86ab4328e1db4abdd5b2195a6e8d87fc4f2b40)