Use uninstall annotation on policies
With the move to being event-driven by default, finalizers were not
being removed during the uninstall process because those policies were
not triggered to be reconciled. Now the `trigger-uninstall` process will
add an annotation to policies with finalizers, which the main controller
now watches for.

The value of the uninstall annotation is now a timestamp, as opposed to
just being set to "true" - this helps provide visibility, and lets the
`trigger-uninstall` process to update the annotation multiple times to
ensure a reconcile is triggered.

Refs:
 - https://issues.redhat.com/browse/ACM-14707

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 29182fcbe55c1cecbf05dacde48d8f263ea57c78)