Send operator policy event if status is reset
ConfigurationPolicies were already handling this, and the predicate used
there was modified for OperatorPolicy so that it is watching for the
same events.

The new test for this behavior was added to the existing CRD status
reporting test, because it benefits from being run after the full
operator has been installed, and adding a separate test that does that
would be an inefficient use of test time.

Refs:
 - https://issues.redhat.com/browse/ACM-15049

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 253fa22bf8596c112ce902e2b617211b743b81c8)