Adopt existing sub when packagemanifest not found
Previously, if the pacakagemanifest was missing, perhaps due to the
operator no longer being available in the catalog, the operator policy
would be marked as invalid, and not report on the potentially helpful
status of the subscription. Now the existing subscription can be used
for the default values, and a more helpful status can be reported.

Refs:
 - https://issues.redhat.com/browse/ACM-14617

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit f2a42b5cd9a91e77c528389db5a4df91c7062a48)