Fix alreadyEvaluated check for dryrun
It appears that the fake API server used by dryrun does not supply
resource versions, so when multiple object templates were supplied in a
ConfigurationPolicy to dryrun, it would incorrectly think that the later
templates had already been evaluated and didn't need to be updated.

Now if the resource version is empty, it will not consider the object as
already evaluated.

Refs:
 - https://issues.redhat.com/browse/ACM-15899

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 1652b317692dee8e1357781f64ec8db59982967b)