Fix mustnothave with ObjectSelector
Previously, the way that names were set on "desired" objects when
ObjectSelector is used was causing a mustnothave policy to consider all
of the objects matching the selector to be noncompliant, as opposed to
only the objects matching other details in the object template.

Now the "desired" objects are filtered in this specific case, in order
to only mark the objects matched by the selector AND the other fields in
the objectDefinition as noncompliant, and to be deleted when the policy
is in enforce mode.

Refs:
 - https://issues.redhat.com/browse/ACM-15772

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 6fb38cebb908c705446f6c85d0b0f3c455e76cf2)